Ignore the Extras folder
This can prevent a catastrophic problem where the Extras folder is deleted on remote servers if the juntion on the regional server disappears. Robo is set to skip junctions, but if the local junction is missing, it doesn't know that the remote folder should be ignored because it use to match a local junction.